+++ src/db_adapters/assets/event_types.rs
@@ -4,9 +4,51 @@ use serde::{Deserialize, Serialize};
 #[serde(tag = "standard")]
 #[serde(rename_all = "snake_case")]
 pub(crate) enum NearEvent {
+    Nep141(Nep141Event),
     Nep171(Nep171Event),
 }
 
+// *** NEP-141 FT ***
+#[derive(Serialize, Deserialize, Debug)]
+pub(crate) struct Nep141Event {
+    pub version: String,
+    #[serde(flatten)]
+    pub event_kind: Nep141EventKind,
+}
+
+#[derive(Serialize, Deserialize, Debug)]
+#[serde(tag = "event", content = "data")]
+#[serde(rename_all = "snake_case")]
+#[allow(clippy::enum_variant_names)]
+pub(crate) enum Nep141EventKind {
+    FtMint(Vec<FtMintData>),
+    FtTransfer(Vec<FtTransferData>),
+    FtBurn(Vec<FtBurnData>),
+}
+
+#[derive(Serialize, Deserialize, Debug)]
+pub(crate) struct FtMintData {
+    pub owner_id: String,
+    pub amount: String,
+    pub memo: Option<String>,
+}
+
+#[derive(Serialize, Deserialize, Debug)]
+pub(crate) struct FtTransferData {
+    pub old_owner_id: String,
+    pub new_owner_id: String,
+    pub amount: String,
+    pub memo: Option<String>,
+}
+
+#[derive(Serialize, Deserialize, Debug)]
+pub(crate) struct FtBurnData {
+    pub owner_id: String,
+    pub amount: String,
+    pub memo: Option<String>,
+}
+
+// *** NEP-171 NFT ***
 #[derive(Serialize, Deserialize, Debug)]
 pub(crate) struct Nep171Event {
     pub version: String,