+++ src/db_adapters/execution_outcomes.rs
@@ -1,25 +1,28 @@
 use actix_diesel::dsl::AsyncRunQueryDsl;
-use diesel::{ExpressionMethods, PgConnection, QueryDsl};
+use cached::Cached;
+use diesel::PgConnection;
+use futures::future::try_join_all;
 
 use crate::models;
 use crate::schema;
-use diesel::pg::expression::array_comparison::any;
 
 pub(crate) async fn store_execution_outcomes(
     pool: &actix_diesel::Database<PgConnection>,
     shards: &[near_indexer::IndexerShard],
     block_timestamp: u64,
+    receipts_cache: crate::ReceiptsCache,
 ) -> anyhow::Result<()> {
-    for shard in shards {
+    let futures = shards.iter().map(|shard| {
         store_execution_outcomes_for_chunk(
             pool,
             &shard.receipt_execution_outcomes,
             shard.shard_id,
             block_timestamp,
+            std::sync::Arc::clone(&receipts_cache),
         )
-        .await?;
-    }
-    Ok(())
+    });
+
+    try_join_all(futures).await.map(|_| ())
 }
 
 /// Saves ExecutionOutcome to database and then saves ExecutionOutcomesReceipts
@@ -28,33 +31,20 @@ pub async fn store_execution_outcomes_for_chunk(
     execution_outcomes: &[near_indexer::IndexerExecutionOutcomeWithReceipt],
     shard_id: near_indexer::near_primitives::types::ShardId,
     block_timestamp: u64,
+    receipts_cache: crate::ReceiptsCache,
 ) -> anyhow::Result<()> {
-    let known_receipt_ids: std::collections::HashSet<String> = crate::await_retry_or_panic!(
-        schema::receipts::table
-            .filter(
-                schema::receipts::dsl::receipt_id.eq(any(execution_outcomes
-                    .iter()
-                    .map(|outcome| outcome.execution_outcome.id.to_string())
-                    .collect::<Vec<_>>())),
-            )
-            .select(schema::receipts::dsl::receipt_id)
-            .load_async::<String>(pool),
-        10,
-        "Parent Receipt for ExecutionOutcome was fetched".to_string(),
-        &execution_outcomes
-    )
-    .unwrap_or_default()
-    .into_iter()
-    .collect();
-
     let mut outcome_models: Vec<models::execution_outcomes::ExecutionOutcome> = vec![];
     let mut outcome_receipt_models: Vec<models::execution_outcomes::ExecutionOutcomeReceipt> =
         vec![];
-    for (index_in_chunk, outcome) in execution_outcomes
-        .iter()
-        .filter(|outcome| known_receipt_ids.contains(&(outcome.execution_outcome.id).to_string()))
-        .enumerate()
-    {
+    let mut receipts_cache_lock = receipts_cache.lock().await;
+    for (index_in_chunk, outcome) in execution_outcomes.iter().enumerate() {
+        // Trying to take the parent Transaction hash for the Receipt from ReceiptsCache
+        // remove it from cache once found as it is not expected to observe the Receipt for
+        // second time
+        let parent_transaction_hash = receipts_cache_lock.cache_remove(
+            &crate::ReceiptOrDataId::ReceiptId(outcome.execution_outcome.id),
+        );
+
         let model = models::execution_outcomes::ExecutionOutcome::from_execution_outcome(
             &outcome.execution_outcome,
             index_in_chunk as i32,
@@ -70,15 +60,27 @@ pub async fn store_execution_outcomes_for_chunk(
                 .receipt_ids
                 .iter()
                 .enumerate()
-                .map(
-                    |(index, receipt_id)| models::execution_outcomes::ExecutionOutcomeReceipt {
+                .map(|(index, receipt_id)| {
+                    // if we have `parent_transaction_hash` from cache, then we put all "produced" Receipt IDs
+                    // as key and `parent_transaction_hash` as value, so the Receipts from one of the next blocks
+                    // could find their parents in cache
+                    if let Some(transaction_hash) = &parent_transaction_hash {
+                        receipts_cache_lock.cache_set(
+                            crate::ReceiptOrDataId::ReceiptId(*receipt_id),
+                            transaction_hash.clone(),
+                        );
+                    }
+                    models::execution_outcomes::ExecutionOutcomeReceipt {
                         executed_receipt_id: outcome.execution_outcome.id.to_string(),
                         index_in_execution_outcome: index as i32,
                         produced_receipt_id: receipt_id.to_string(),
-                    },
-                ),
+                    }
+                }),
         );
     }
+
+    // releasing the lock
+    drop(receipts_cache_lock);
 
     crate::await_retry_or_panic!(
         diesel::insert_into(schema::execution_outcomes::table)