+++ src/db_adapters/mod.rs
@@ -9,3 +9,5 @@ pub(crate) mod execution_outcomes;
 pub(crate) mod genesis;
 pub(crate) mod receipts;
 pub(crate) mod transactions;
+
+const CHUNK_SIZE_FOR_BATCH_INSERT: usize = 500;