+++ src/main.rs
@@ -4,10 +4,12 @@ use std::convert::TryFrom;
 extern crate diesel;
 
 use actix_diesel::Database;
+pub use cached::SizedCache;
 use diesel::PgConnection;
+use futures::future::try_join_all;
 use futures::{try_join, StreamExt};
-use tokio::sync::mpsc;
-use tracing::{info, warn};
+use tokio::sync::{mpsc, Mutex};
+use tracing::{debug, info, warn};
 use tracing_subscriber::EnvFilter;
 
 use crate::configs::{Opts, SubCommand};
@@ -27,11 +29,30 @@ const AGGREGATED: &str = "aggregated";
 const INTERVAL: std::time::Duration = std::time::Duration::from_millis(100);
 const MAX_DELAY_TIME: std::time::Duration = std::time::Duration::from_secs(120);
 
+#[derive(Clone, Hash, PartialEq, Eq, Debug)]
+pub enum ReceiptOrDataId {
+    ReceiptId(near_indexer::near_primitives::hash::CryptoHash),
+    DataId(near_indexer::near_primitives::hash::CryptoHash),
+}
+// Creating type aliases to make HashMap types for cache more explicit
+pub type ParentTransactionHashString = String;
+// Introducing a simple cache for Receipts to find their parent Transactions without
+// touching the database
+// The key is ReceiptID
+// The value is TransactionHash (the very parent of the Receipt)
+pub type ReceiptsCache =
+    std::sync::Arc<Mutex<SizedCache<ReceiptOrDataId, ParentTransactionHashString>>>;
+
 async fn handle_message(
     pool: &actix_diesel::Database<PgConnection>,
     streamer_message: near_indexer::StreamerMessage,
     strict_mode: bool,
+    receipts_cache: ReceiptsCache,
 ) -> anyhow::Result<()> {
+    debug!(
+        target: INDEXER_FOR_EXPLORER,
+        "ReceiptsCache #{} \n {:#?}", streamer_message.block.header.height, &receipts_cache
+    );
     db_adapters::blocks::store_block(pool, &streamer_message.block).await?;
 
     // Chunks
@@ -49,6 +70,7 @@ async fn handle_message(
         &streamer_message.block.header.hash,
         streamer_message.block.header.timestamp,
         streamer_message.block.header.height,
+        std::sync::Arc::clone(&receipts_cache),
     );
 
     // Receipts
@@ -58,6 +80,7 @@ async fn handle_message(
         &streamer_message.block.header.hash,
         streamer_message.block.header.timestamp,
         strict_mode,
+        std::sync::Arc::clone(&receipts_cache),
     );
 
     // We can process transactions and receipts in parallel
@@ -72,37 +95,37 @@ async fn handle_message(
         pool,
         &streamer_message.shards,
         streamer_message.block.header.timestamp,
+        std::sync::Arc::clone(&receipts_cache),
     );
 
     // Accounts
     let accounts_future = async {
-        for shard in &streamer_message.shards {
+        let futures = streamer_message.shards.iter().map(|shard| {
             db_adapters::accounts::handle_accounts(
                 pool,
                 &shard.receipt_execution_outcomes,
                 streamer_message.block.header.height,
             )
-            .await?;
-        }
-        Ok(())
+        });
+
+        try_join_all(futures).await.map(|_| ())
     };
 
-    // Assets (NFT)
-    let nft_events_future =
-        db_adapters::assets::non_fungible_token_events::store_nft(pool, &streamer_message);
+    // Event-based entities (FT, NFT)
+    let assets_events_future = db_adapters::assets::events::store_events(pool, &streamer_message);
 
     if strict_mode {
         // AccessKeys
         let access_keys_future = async {
-            for shard in &streamer_message.shards {
+            let futures = streamer_message.shards.iter().map(|shard| {
                 db_adapters::access_keys::handle_access_keys(
                     pool,
                     &shard.receipt_execution_outcomes,
                     streamer_message.block.header.height,
                 )
-                .await?;
-            }
-            Ok(())
+            });
+
+            try_join_all(futures).await.map(|_| ())
         };
 
         // StateChange related to Account
@@ -117,14 +140,14 @@ async fn handle_message(
             execution_outcomes_future,
             accounts_future,
             access_keys_future,
-            nft_events_future,
+            assets_events_future,
             account_changes_future,
         )?;
     } else {
         try_join!(
             execution_outcomes_future,
             accounts_future,
-            nft_events_future
+            assets_events_future
         )?;
     }
     Ok(())
@@ -150,13 +173,27 @@ async fn listen_blocks(
         );
     }
     info!(target: crate::INDEXER_FOR_EXPLORER, "Stream has started");
+
+    // We want to prevent unnecessary SELECT queries to the database to find
+    // the Transaction hash for the Receipt.
+    // Later we need to find the Receipt which is a parent to underlying Receipts.
+    // Receipt ID will of the child will be stored as key and parent Transaction hash/Receipt ID
+    // will be stored as a value
+    let receipts_cache: ReceiptsCache =
+        std::sync::Arc::new(Mutex::new(SizedCache::with_size(100_000)));
+
     let handle_messages =
         tokio_stream::wrappers::ReceiverStream::new(stream).map(|streamer_message| {
             info!(
                 target: crate::INDEXER_FOR_EXPLORER,
                 "Block height {}", &streamer_message.block.header.height
             );
-            handle_message(&pool, streamer_message, strict_mode)
+            handle_message(
+                &pool,
+                streamer_message,
+                strict_mode,
+                std::sync::Arc::clone(&receipts_cache),
+            )
         });
     let mut handle_messages = if let Some(stop_after_n_blocks) = stop_after_number_of_blocks {
         handle_messages
@@ -244,9 +281,25 @@ fn main() {
     // Indexer should fail if .env file with credentials is missing/wrong
     let pool = models::establish_connection();
 
+    let opts: Opts = Opts::parse();
+
     let mut env_filter = EnvFilter::new(
-        "tokio_reactor=info,near=info,stats=info,telemetry=info,indexer=info,indexer_for_explorer=info,aggregated=info",
+        "tokio_reactor=info,near=info,stats=info,telemetry=info,indexer=info,aggregated=info",
     );
+
+    if opts.debug {
+        env_filter = env_filter.add_directive(
+            "indexer_for_explorer=debug"
+                .parse()
+                .expect("Failed to parse directive"),
+        );
+    } else {
+        env_filter = env_filter.add_directive(
+            "indexer_for_explorer=info"
+                .parse()
+                .expect("Failed to parse directive"),
+        );
+    };
 
     if let Ok(rust_log) = std::env::var("RUST_LOG") {
         if !rust_log.is_empty() {
@@ -266,8 +319,6 @@ fn main() {
         .with_env_filter(env_filter)
         .with_writer(std::io::stderr)
         .init();
-
-    let opts: Opts = Opts::parse();
 
     let home_dir = opts.home_dir.unwrap_or_else(near_indexer::get_default_home);
 
@@ -327,6 +378,7 @@ fn main() {
             config.download_config_url.as_ref().map(AsRef::as_ref),
             config.boot_nodes.as_ref().map(AsRef::as_ref),
             config.max_gas_burnt_view,
-        ),
+        )
+        .expect("Failed to initiate config"),
     }
 }